+++ CHANGES.rst
@@ -7,6 +7,7 @@ Version 1.2.0
 Unreleased
 
 - cachelib is now used as backend. PR `#308 <https://github.com/pallets-eco/flask-caching/pull/308>`_.
+- A ``DynamoDbCache`` backend has been add to the user contributed backends.
 
 
 Version 1.10.1